Counter application with background task and local notification
Start counter, background task of counting up is triggered if the app goes into background and ended when the app returns to foreground (the app resumes counting until Stop button pressed). While the app is in background, the user is kept informed by notifications each time the count reaches a multiple of 10.